ASDPLNG-117 Edit ignores for POSTROUTING:nat:IPv4
Rules in POSTROUTING:nat:IPv4 on LSST puppet server were getting
removed causing the container to lose network access.

The current list of ignores for that table+chain are not generic enough.
It will ignore rules that match "172.17", "172.18", "172.19"
but LSST was getting rules with 172.21, and there is no immediate
knowledge of why or how docker chooses these addresses

So need to loosen the matching so it will ignore any rule that matches
any of these regexes:

  POSTROUTING:nat:IPv4:
    - '-s 172\.'  # yamllint disable-line rule:quoted-strings
    - '-d 172\.'  # yamllint disable-line rule:quoted-strings

Need to add the string in single quotes otherwise puppet will complain
as it tries to interpret the escape character and you get an error.
Our .yamllint.yaml sets strings to use double-quotes, so use an in-line
flag to turn off that check, otherwise would get a yamllint error